Correct privilege granting powershell script.
Originally worked only by coincidence due to a memory related coding error. It
would fail on the first execution with a Win32Exception saying "The parameter is
incorrect" referring to the username parameter. In a command prompt, it would
seem to work as well. The [code was sourced from stackoverflow](http://stackoverflow.com/questions/26392151/enabling-a-local-user-right-assignment-in-powershell/26393118#26393118)
but it shows up in various other places with minor differences.

The [corrections for the memory related coding error are sourced from stackoverflow](http://stackoverflow.com/questions/1147914/why-might-lsaaddaccountrights-return-status-invalid-parameter/14469248#14469248) too.
It resolves the memory error.