Use a custom storesymbols.bat.
Until ansible provides better support for windows (e.g. lineinfile), use
a custom bat file with the desired change.

The actual desired outcome was not to touch the utility but to provide the z:
mapping that it uses. Unfortunately, Windows makes this impossible (or so
obscure and difficult that discovering how to do it has been elusive) without
manual intervention thus making automation also impossible. Even if it were
possible, this workaround is faster and more effective than wasting further time
on Windows.